Hw5: rename package dao -> domain
