+++ scrapper/src/main/java/edu/java/domain/LinkRepository.java
@@ -1,4 +1,4 @@
-package edu.java.dao;
+package edu.java.domain;
 
 import edu.java.dto.dao.LinkDto;
 import java.time.Duration;
